requirements.txt
numpy
pandas
matplotlib
torch
torchvision
torch_optimizer
albumentations
matplotlib
opencv-python
Pillow
res-mlp-pytorch
pgmpy
seaborn